Split up ApiV2Controller
